==> 01-data-manipulation-scripts/02-data-cleaning/08-nutrients/02-clean-calcofi.Rmd <==
---
title: "calCOFI"
author: "Elke Windschitl"
date: "2023-03-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sf)
library(tmap)
```

```{r}
#---- Set up the data directory (wherever your download of our Google Shared Drive lives)
data_dir <- "/capstone/kelpgeomod/google_drive_download"
```

```{r}
# Read in AOI
aoi <- st_read(file.path(data_dir, "02-intermediate-data/02-aoi-sbchannel-shapes-intermediate/aoi-sbchannel.shp"))

# Read in CalCOFI cast data
CalCOFI_cast <- read_csv(file.path(data_dir, "01-raw-data/09-nutrients-various-raw/02-nutrients-calcofi-raw/194903-202001_Cast.csv")) %>% 
  filter(Year >= 2014) %>% 
  dplyr::select(c("Cst_Cnt", "Quarter", "Year", "Lat_Dec", "Lon_Dec"))

# Read in the CalCOFI bottle data
CalCOFI_bottle <- read_csv(file.path(data_dir, "01-raw-data/09-nutrients-various-raw/02-nutrients-calcofi-raw/194903-202001_Bottle.csv"), locale = locale(encoding = "latin1"))

# Join cast and bottle data + vectorize lat lon point data and put into our nutrient data format
CalCOFI_join <- CalCOFI_bottle %>% 
  right_join(CalCOFI_cast, by = "Cst_Cnt") %>% 
  dplyr::select(c("Lon_Dec", "Lat_Dec", "Depthm", "T_degC", "Quarter", "Year","PO4uM", "NO2uM", "NO3uM", "NH3uM")) %>% 
  filter(Depthm <= 10) %>% 
  st_as_sf(coords = c(lon = "Lon_Dec", lat = "Lat_Dec"), crs = "EPSG: 4326") %>% 
  st_filter(aoi) %>% 
  relocate("geometry", .after = "Year") %>% 
  relocate("NO3uM", .after = "geometry") %>% 
  relocate("NO2uM", .after = "NO3uM") %>% 
  relocate("NH3uM", .after = "NO2uM") %>% 
  set_names(c("depth", "temp", "quarter", "year", "geometry", "nitrate", "nitrite", "ammonia", "phosphate"))
```

```{r}
# Get averages of points in one location
CalCOFI_ave <- CalCOFI_join %>% 
  group_by(quarter, year, geometry) %>% 
  summarise(temp = mean(temp, na.rm = TRUE),
            nitrate = mean(nitrate, na.rm = TRUE),
            nitrite = mean(nitrite, na.rm = TRUE),
            ammonia = mean(ammonia, na.rm = TRUE),
            phosphate = mean(phosphate, na.rm = TRUE)) %>% 
  mutate("source" = "CalCOFI") 

# Check it out
tm_shape(CalCOFI_ave) +
  tm_dots()

# Remove geometry for export to match our nutrient data structure
cal_unlist <- CalCOFI_ave %>%
  st_coordinates() %>%
  as.data.frame() %>%
  setNames(c("lon", "lat")) %>%
  bind_cols(CalCOFI_ave) %>% 
  dplyr::select(-geometry) %>%
  mutate("nitrate_nitrite" = nitrate + nitrite, .after = nitrite) %>% # create nitrate_nitrite combined column
  rename("ammonium" = "ammonia") # Natalie said the naming difference is likely do to how each data source quantified the measurement but it was okay to combine them. 
```

```{r}
#---- Write to data files
# Change this file path to your own to re-write data
write_csv(x = cal_unlist, file = "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/unjoined/calcofi.csv")
```
