==> 01-data-manipulation-scripts/02-data-cleaning/08-nutrients/03-clean-lter-bottle.Rmd <==
---
title: "Kelp LTER Bottle Data Wrangling"
author: "Erika Egg"
date: '`r format(Sys.time(), "%m/%d/%Y")`'
output: html_document
---

```{r}
#---- Load in packages
library(tidyverse)
library(janitor)
library(sf)
library(tmap)
library(lubridate)

#---- Set up the data directory (wherever your download of our Google Shared Drive lives)
data_dir <- "/capstone/kelpgeomod/google_drive_download"
```

```{r}
# Read in the file of verified sites. Data link: https://sbclter.msi.ucsb.edu/data/catalog/package/?package=knb-lter-sbc.10
lter_bottle_data <- read.table(file.path(data_dir, "01-raw-data/09-nutrients-various-raw/03-nutrients-lter-bottle-raw/LTER_monthly_bottledata_20220930.txt"), sep = ";", header = TRUE)
```

#
```{r}
# Clean the column names 
data_cleaned <- lter_bottle_data %>% 
  clean_names()

data_selected <- data_cleaned %>%
  mutate("latitude_n" = as.numeric(latitude_n),
         "longitude_e" = as.numeric(longitude_e),
         "yyyy_mm_dd" = as.Date(yyyy_mm_dd)) %>%
  dplyr::select(date = yyyy_mm_dd, lon = longitude_e, lat = latitude_n, depth = target_depth_m, phosphate = po4_umol_l, nitrate_nitrite = no2_no3_umol_l, ammonium = ammonia_umol_l, temp = ta_pco2_temperature_deg_c)

# Get month dates to just be first of month
data_selected$date <- floor_date(data_selected$date, unit = "month")

# We also want to set what our latitude and longitude bounds of our area of interest are.
lower_lat <- 33.85
upper_lat <- 34.59
lower_lon <- -120.65
upper_lon <- -118.80

# Filter to relevant time period and AOI, creating quarter and year columns and standardizing missing values
data_2014_2022 <- data_selected %>%
  filter(year(date) >= 2014) %>% # get data for our time period of interest
  mutate(quarter = quarter(date),
         year = year(date)) %>% # get year and quarter columns from the date
  mutate_at(vars(phosphate, nitrate_nitrite, ammonium, temp), ~ifelse(. %in% "99999", NA, .)) %>% # change 99999 missing value code to NA
  mutate_at(vars(phosphate, nitrate_nitrite, ammonium), ~ifelse(. %in% "BDL", 0, .)) %>% # change BDL missing value code to 0
  mutate(phosphate = as.numeric(phosphate),
         nitrate_nitrite = as.numeric(nitrate_nitrite),
         ammonium = as.numeric(ammonium)) %>% # ensure they are numeric for later means
  filter(lat >= lower_lat & lat <= upper_lat & lon >= lower_lon & lon <= upper_lon) # filter to our area of interest

# Summarize data by quarter
lter_bottle_data_standardized <- data_2014_2022 %>%
  group_by(lat, lon, year, quarter) %>% # add depth here if want to keep depth as a dividing factor
  summarize(temp = mean(temp, na.rm = TRUE),
            phosphate = mean(phosphate, na.rm = TRUE),
            nitrate_nitrite = mean(nitrate_nitrite, na.rm = TRUE),
            ammonium = mean(ammonium, na.rm = TRUE)) # take quarterly means of monthly data

lter_bottle_data_standardized['source'] = 'lter_bottle' # add a source column for tracking when combining with other nutrient data
```


```{r}
# Sanity check to ensure it's within the AOI: convert the lat and lon columns to point geometry in order to mask the data
# However, based on the above code, we should already be good
# data_sf <- lter_bottle_data_standardized %>% 
#   st_as_sf(coords = c("lon", "lat"), 
#            crs = 4326) 
# 
# # Load the mask
# mask <- raster::raster(file.path(data_dir, "02-intermediate-data/03-mask-intermediate/main-mask.tif"))
# 
# # Make sure the sf data has the same CRS as the mask
# data_sf_transformed <- st_transform(data_sf, crs = raster::projection(mask))
# 
# # Crop the sf data to the extent of the mask
# data_sf_cropped <- st_crop(data_sf_transformed, mask)

# In this case, data_sf_cropped is equivalent to lter_bottle_data_standardized, except has a geometry column instead of lat/lon
# I will continue with lter_bottle_data_standardized

# However, the below code will convert your geometry back to lat/lon if you run this code and determine your data changes with it
# data_lat_lon <- data_sf_cropped %>%
#   st_coordinates() %>%
#   as.data.frame() %>%
#   setNames(c("lat", "lon")) %>%
#   bind_cols(data_sf_cropped) %>%
#   select(-geometry)

# Write csv
write.csv(lter_bottle_data_standardized, file = "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/unjoined/lter-bottle.csv")
# saveRDS(lter_bottle_data_standardized, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/unjoined/lter-bottle.csv") # if an RDS version is desired
```
