==> 01-data-manipulation-scripts/02-data-cleaning/08-nutrients/04-clean-lter-nitrate.Rmd <==
---
title: "nitrate_lter"
author: "Elke Windschitl"
date: "2023-03-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sf)
library(terra)
library(lubridate)
library(tmap)
```

```{r}
#---- Set up the data directory (wherever your download of our Google Shared Drive lives)
data_dir <- "/capstone/kelpgeomod/google_drive_download"
```

```{r}
# Read in AOI
aoi <- st_read(file.path(data_dir, "02-intermediate-data/02-aoi-sbchannel-shapes-intermediate/aoi-sbchannel.shp"))

# Load nitrogen data
nitrate_raw <- read_csv(file.path(data_dir, "01-raw-data/09-nutrients-various-raw/04-nutrients-lter-nitrate-raw/kelp_no3_waves_quarterly_long.csv"))

# Load location data for each site
locations_raw <- read_csv(file.path(data_dir, "01-raw-data/09-nutrients-various-raw/04-nutrients-lter-nitrate-raw/locations.csv"))
```

```{r}
# Select for columns of interest for this data set 
nitrate_dat <- nitrate_raw %>% 
  dplyr::select(c("site_id", "year", "quarter", "no3")) %>% 
  filter(year >= 2014 & year <= 2022) %>% 
  left_join(locations_raw, by = "site_id") #%>% # left join location data to the data set
```

```{r}
# Vectorize lat lon point data and put into our nutrient data format
nitrate_geo <- nitrate_dat %>% 
  st_as_sf(coords = c("lon", "lat")) %>% 
  dplyr::select(-site_id) %>% 
  rename("nitrate" = "no3") %>% 
  relocate("quarter", .before = "year") %>% 
  relocate("geometry", .after = "year") %>% 
  mutate("source" = "lter_nitrate")
st_crs(nitrate_geo) = 4326
nitrate_geo <- nitrate_geo %>% 
  st_filter(aoi)

# Check it out
tmap_mode("view")
tm_shape(nitrate_geo) +
  tm_dots()

# Remove geometry for export to match our nutrient data structure
lter_unlist <- nitrate_geo %>%
  st_coordinates() %>%
  as.data.frame() %>%
  setNames(c("lon", "lat")) %>%
  bind_cols(nitrate_geo) %>% 
  select(-geometry)
```

```{r}
#---- Write to data files
# Change this file path to your own to re-write data
write_csv(x = lter_unlist, file = "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/unjoined/lter-nitrate.csv")
```
