==> 01-data-manipulation-scripts/02-data-cleaning/08-nutrients/06-clean-usgs.Rmd <==
---
title: "usgs"
author: "Javier Patron"
date: "2023-03-28"
---

### **Assumptions and points to consider**
1.  The -9999.00 values are no good for phosphate, ammonium, nitrate_and_nitrite in point conception, and Gill Islands 2
2.  The final values per quarter are the mean of the 10 observations per location
im using the function rm.na = T, for NA values 

Load the Libraries

```{r}
library(tidyverse)
library(janitor)
library(raster)
library(sp)
library(tmap)
library(sf)
library(stars)
library(dplyr)
library(terra)
library(lubridate)

```


Set the kelpGeoMod Google Drive Working Directory
```{r}
#---- Set up the data directory (wherever your download of our Google Shared Drive lives)
drive_dir <- "/capstone/kelpgeomod/google_drive_download"
```


Read the data
```{r}
         
data_version2 <- read_csv(file.path(drive_dir, "01-raw-data/09-nutrients-various-raw/06-nutrients-usgs-bottle-raw/RL-19-05_BTL_CTD_data_v2.0.csv"))
data_version3 <- read_csv(file.path(drive_dir, "01-raw-data/09-nutrients-various-raw/06-nutrients-usgs-bottle-raw/SH-18-12_BTL_CTD_data_v3.0.csv"))

```

Explore the data
```{r}
# convert lat and long columns to sf
sf_data_2 <- data_version2 |> 
  st_as_sf(coords = c("Longitude", "Latitude"),
           crs = 4326)

# convert lat and long columns to sf
sf_data_3 <- data_version3 |> 
  st_as_sf(coords = c("Longitude", "Latitude"),
           crs = 4326)

```

```{r}
# Define the rectangle as a polygon
rectangle <- st_polygon(list(rbind(c(-120.65, 34.59), 
                                   c(-120.65, 33.85), 
                                   c(-118.80, 33.85), 
                                   c(-118.80, 34.59), 
                                   c(-120.65, 34.59)))) # You need to close the polygon with rewriting the last point

# Convert the polygon to an sf object
area_sf <- st_sf(geometry = st_sfc(rectangle))
area_sf <- st_set_crs(area_sf, 4326)

```

TESTING: See where the data points are
```{r}
# Create map to see where the samples were taken
tmap_mode(mode = "view")

tm_shape(sf_data_2) +
  tm_dots(col = "green") +

tm_shape(sf_data_3) +
  tm_dots(col = "blue") +

tm_shape(area_sf) +
  tm_borders(lwd = 1, col = "red") 
```

## Data Cleaning (Read, clean, select the columns and filter the sites)
```{r}
data2 <- read_csv(file.path(drive_dir, "01-raw-data/09-nutrients-various-raw/06-nutrients-usgs-bottle-raw/RL-19-05_BTL_CTD_data_v2.0.csv")) |> 
  clean_names() |> 
  dplyr::select(c("site_name","date", "temperature", "phosphate", "ammonium", "nitrate_and_nitrite", "latitude", "longitude")) |> 
  filter(site_name == "Annapaca")

data3 <- read_csv(file.path(drive_dir, "01-raw-data/09-nutrients-various-raw/06-nutrients-usgs-bottle-raw/SH-18-12_BTL_CTD_data_v3.0.csv")) |> 
  clean_names() |> 
  dplyr::select(c("site_name", "date", "temperature", "phosphate", "ammonium", "nitrate_and_nitrite", "latitude", "longitude")) |> 
  filter(site_name  %in% c("Gull Island", "Anacapa"))

# Join both datasets
joined_data <- rbind(data2, data3)

# Convert -9999.00 values to NA
joined_data[joined_data == -9999.00] <- NA

#Convert date as lubridate
joined_data$date <- as.Date(joined_data$date, format = "%m/%d/%y")


# create a new column that has year and quarter, and select columns of interest
joined_data <- joined_data %>% 
  mutate(year = year(date),
         quarter = quarter(date)) %>% 
  dplyr::select(year, quarter, temperature, nitrate_and_nitrite, phosphate, ammonium, latitude, longitude) %>% 
  mutate(ammonium = ifelse(ammonium == "<0.019", 0.020, as.numeric(ammonium))) %>% 
  rename(lat = latitude) %>% 
  rename(lon = longitude)


# group_by and create the mean
final_joined_data <- joined_data %>%
  group_by(year, quarter, lat, lon) %>%
  summarise(temp = mean(temperature, na.rm = TRUE),
            nitrate_nitrite = mean(nitrate_and_nitrite, na.rm = TRUE),
            phosphate = mean(phosphate, na.rm = TRUE),
            ammonium = mean(ammonium, na.rm = TRUE))

final_joined_data['source'] = 'USGS' # add a source column for tracking when combining with other nutrient data
```


```{r}
# Write spatial object to RDS file
write_csv(final_joined_data, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/unjoined/usgs.csv")
```
