==> 01-data-manipulation-scripts/02-data-cleaning/08-nutrients/07-join-nutrients.Rmd <==
---
title: "Nutrient Join"
author: "Jessica French"
date: "`r Sys.Date()`"
output: html_document
---
## Purpose

The purpose of this notebook is to join the intermediate nutrient data into a single dataframe and create raster bricks of nutrients with one layer per year and quarter. 

## Set-up

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r label = "read in packages"}
library(tidyverse)
library(tmap)
library(sf)
library(naniar)
library(here)
library(terra)
library(raster)
```

```{r label = "define data directory"}
data_dir <- "/capstone/kelpgeomod/google_drive_download"
```

## Read in data sets

data sets to read in
- Plumes and Blooms
- CalCOFI
- USGS
- LTER Bottle
- LTER Nitrate

```{r label = "Read in Data"}
# Read in data from tsosie

# Plumes and Blooms
PB <- read_csv(file.path(data_dir, "02-intermediate-data/09-nutrients-intermediate/unjoined/plumes-blooms.csv")) |> 
  st_as_sf(coords = c("lon", "lat"))

# CalCOFI
calCOFI <- read_csv(file.path(data_dir, "02-intermediate-data/09-nutrients-intermediate/unjoined/calcofi.csv")) |> 
  st_as_sf(coords = c("lon", "lat"))

# USGS
USGS <- read_csv(file.path(data_dir, "02-intermediate-data/09-nutrients-intermediate/unjoined/usgs.csv")) |> 
  st_as_sf(coords = c("lon", "lat"))

# LTER Nitrate
lter_nitrate <- read_csv(file.path(data_dir, "02-intermediate-data/09-nutrients-intermediate/unjoined/lter-nitrate.csv")) |> 
  st_as_sf(coords = c("lon", "lat"))

# LTER bottle 
lter_bottle <- read_csv(file.path(data_dir, "02-intermediate-data/09-nutrients-intermediate/unjoined/lter-bottle.csv")) |> 
  st_as_sf(coords = c("lon", "lat"))
```

```{r label = "combine"}
# Combine data sets
# Create list of data sets
data_set_list <- list(calCOFI, PB, USGS, lter_bottle, lter_nitrate)

# test with two that have different columns
joined_nutrients <- bind_rows(data_set_list) |> 
  replace_with_na(replace = list(nitrate = "NaN", 
                                 nitrite = "NaN", 
                                 nitrate_nitrite = "NaN", 
                                 phosphate = "NaN", 
                                 ammonium = "NaN", 
                                 temp = "NaN"))

```


```{r label = "read in mask"}
# read in mask
mask <- rast(file.path(data_dir, "02-intermediate-data/03-mask-intermediate/main-mask.tif"))
```

## Create year and quarter raster bricks

```{r label = "source in functions"}
# source in function that creates a list of data frames by quarter
source(here("01-data-manipulation-scripts/03-data-interpolation/02-functions/02-prep.R"))
source(here("01-data-manipulation-scripts/03-data-interpolation/02-functions/03-implement.R"))
```

```{r label = "phosphate quarters"}
# no need to assign this to anything, assignment is done in the function. 
# use year and quarter if interpolating for each year and quarter. 
idw_prep_year_quarters("phosphate", joined_nutrients)
idw_prep_year_quarters("nitrate_nitrite", joined_nutrients)
idw_prep_year_quarters("ammonium", joined_nutrients)
```

```{r label = "rasterize year quarter nutrients"}
# use idw_implement_rasterize.R function to convert each year and quarter to a raster. 

# phosphate ----
phosphate_year_quarter_raster_all <- idw_implement_rasterize(phosphate_year_quarter_df_list, 
                                                         "phosphate", 
                                                         mask, 
                                                         all = TRUE)

# save phosphat_year_quarter_all as raster brick and save as tif for pivoted data frame. 
phosphate_raster_list <- lapply(seq_along(phosphate_year_quarter_raster_all), function(i) {
  raster(phosphate_year_quarter_raster_all[[i]])
})

phosphate_idw_brick <- raster::brick(phosphate_raster_list)

#nitrate_nitrite ----
nitrate_year_quarter_raster_all <- idw_implement_rasterize(nitrate_nitrite_year_quarter_df_list, 
                                                         "nitrate_nitrite", 
                                                         mask, 
                                                         all = TRUE)

# concert to Raster object so can be saved as a brick
nitrate_list <- lapply(seq_along(nitrate_year_quarter_raster_all), function(i) {
  raster(nitrate_year_quarter_raster_all[[i]])
})

nitrate_nitrite_idw_brick <- raster::brick(nitrate_list)

# ammonium ----
ammonium_year_quarter_raster_all <- idw_implement_rasterize(ammonium_year_quarter_df_list, 
                                                         "ammonium", 
                                                         mask, 
                                                         all = TRUE)

# save ammonium_year_quarter_all as raster brick and save as tif for pivoted data frame. 
ammonium_list <- lapply(seq_along(ammonium_year_quarter_raster_all), function(i) {
  raster(ammonium_year_quarter_raster_all[[i]])
})

ammonium_idw_brick <- raster::brick(nitrate_list)
```

```{r label = "save year quarter raster bricks"}
# write each brick to file 
writeRaster(phosphate_idw_brick, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/joined/phosphate-brick.tif", format = "GTiff", overwrite = TRUE)

writeRaster(nitrate_nitrite_idw_brick, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/joined/nitrate-nitrite-brick.tif", format = "GTiff", overwrite = TRUE)

writeRaster(ammonium_idw_brick, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/joined/ammonium-brick.tif", format = "GTiff", overwrite = TRUE)

# write each brick to file 
writeRaster(phosphate_idw_brick, "/capstone/kelpgeomod/google_drive_download/04-deliverables/02-full-synthesized/phosphate-brick.tif", format = "GTiff", overwrite = TRUE)

writeRaster(nitrate_nitrite_idw_brick, "/capstone/kelpgeomod/google_drive_download/04-deliverables/02-full-synthesized/nitrate-nitrite-brick.tif", format = "GTiff", overwrite = TRUE)

writeRaster(ammonium_idw_brick, "/capstone/kelpgeomod/google_drive_download/04-deliverables/02-full-synthesized/ammonium-brick.tif", format = "GTiff", overwrite = TRUE)
```

```{r label = "write to csv"}
# Remove geometries for joining process
joined_nutrients2 <- joined_nutrients %>% 
  st_coordinates() %>% 
  as.data.frame() %>% 
  setNames(c("lon", "lat")) %>% 
  bind_cols(joined_nutrients) %>% 
  dplyr::select(-geometry) %>% 
  dplyr::select(-...13)

# write to csv
write_csv(joined_nutrients2, "/capstone/kelpgeomod/google_drive_download/02-intermediate-data/09-nutrients-intermediate/joined/joined-nutrients.csv")
```
